+++ test/libsolidity/semanticTests/array/copying/array_copy_storage_storage_struct.sol
@@ -17,7 +17,7 @@ contract c {
 }
 // ----
 // test() -> 4, 5
-// gas irOptimized: 238692
+// gas irOptimized: 238623
 // gas legacy: 238736
 // gas legacyOptimized: 237159
 // storageEmpty -> 1